+++ admin/src/scenes/plan-transport/util.js
@@ -3,6 +3,12 @@ import Avion from "../../assets/icons/Avion";
 import Bus from "../../assets/icons/Bus";
 import Fusee from "../../assets/icons/Fusee";
 import Train from "./ligne-bus/components/Icons/Train";
+import { ES_NO_LIMIT, ROLES, formatDateFR, formatDateFRTimezoneUTC, translate } from "snu-lib";
+import FileSaver from "file-saver";
+import { toastr } from "react-redux-toastr";
+import dayjs from "dayjs";
+import API from "../../services/api";
+import * as XLSX from "xlsx";
 
 export function formatRate(value, total, fractionDigit = 0, allowMoreThan100 = false) {
   if (total === 0 || total === undefined || total === null || value === undefined || value === null) {
@@ -64,7 +70,7 @@ export const GROUPSTEPS = {
   CONFIRM_DELETE_GROUP: "CONFIRM_DELETE_GROUP",
 };
 
-export function getRegionsFromBusLine(line) {
+function getRegionsFromBusLine(line) {
   let regions = [];
   regions.push(line.centerRegion);
   for (const pdr of line.pointDeRassemblements) {
@@ -73,11 +79,214 @@ export function getRegionsFromBusLine(line) {
   return regions;
 }
 
-export function getDepartmentsFromBusLine(line) {
+function getDepartmentsFromBusLine(line) {
   let departments = [];
   departments.push(line.centerDepartment);
   for (const pdr of line.pointDeRassemblements) {
     departments.push(pdr.department);
   }
   return departments;
+}
+
+function filterBusLinesByRole(lines, user) {
+  const linesWithGeography = lines.map((e) => {
+    return { ...e, regions: getRegionsFromBusLine(e), departments: getDepartmentsFromBusLine(e) };
+  });
+
+  switch (user.role) {
+    case ROLES.ADMIN:
+    case ROLES.TRANSPORTER:
+      return linesWithGeography;
+    case ROLES.REFERENT_DEPARTMENT:
+      return linesWithGeography.filter((line) => user.department.some((dep) => line.departments.includes(dep)));
+    case ROLES.REFERENT_REGION:
+      return linesWithGeography.filter((line) => line.regions.includes(user.region));
+    default:
+      return [];
+  }
+}
+
+export const exportLigneBus = async (user, cohort) => {
+  try {
+    const body = {
+      query: {
+        bool: {
+          must: [{ match_all: {} }, { term: { "cohort.keyword": cohort } }],
+        },
+      },
+    };
+
+    const data = await getAllResults("plandetransport", body);
+    const ligneBus = filterBusLinesByRole(data, user);
+    const ligneIds = ligneBus.map((e) => e._id);
+
+    let meetingPoints = [];
+    for (const line of ligneBus) {
+      for (const mp of line.pointDeRassemblements) {
+        meetingPoints.push(mp);
+      }
+    }
+
+    const esYoungByLine = async () => {
+      let body = {
+        query: {
+          bool: {
+            must: [],
+            filter: [
+              { match_all: {} },
+              { term: { "cohort.keyword": cohort } },
+              { term: { "status.keyword": "VALIDATED" } },
+              { terms: { "ligneId.keyword": ligneIds } },
+              { term: { "cohesionStayPresence.keyword": "true" } },
+              { term: { "departInform.keyword": "false" } },
+            ],
+          },
+        },
+        aggs: {
+          group_by_bus: {
+            terms: {
+              field: "ligneId.keyword",
+              size: ES_NO_LIMIT,
+            },
+          },
+        },
+        track_total_hits: true,
+      };
+
+      return await getAllResults("young-having-meeting-point-in-geography", body);
+    };
+
+    const youngs = await esYoungByLine();
+    console.log("🚀 ~ file: List.js:238 ~ exportDataTransport ~ youngs:", youngs);
+
+    let result = {
+      noMeetingPoint: {
+        youngs: [],
+        meetingPoint: [],
+      },
+      transportInfoGivenByLocal: {
+        youngs: [],
+        meetingPoint: [],
+      },
+    };
+
+    for (const young of youngs) {
+      const tempYoung = {
+        _id: young._id,
+        cohort: young.cohort,
+        firstName: young.firstName,
+        lastName: young.lastName,
+        email: young.email,
+        phone: young.phone,
+        address: young.address,
+        zip: young.zip,
+        city: young.city,
+        department: young.department,
+        region: young.region,
+        birthdateAt: young.birthdateAt,
+        gender: young.gender,
+        parent1FirstName: young.parent1FirstName,
+        parent1LastName: young.parent1LastName,
+        parent1Email: young.parent1Email,
+        parent1Phone: young.parent1Phone,
+        parent1Status: young.parent1Status,
+        parent2FirstName: young.parent2FirstName,
+        parent2LastName: young.parent2LastName,
+        parent2Email: young.parent2Email,
+        parent2Phone: young.parent2Phone,
+        parent2Status: young.parent2Status,
+        statusPhase1: young.statusPhase1,
+        meetingPointId: young.meetingPointId,
+        ligneId: young.ligneId,
+      };
+      if (young.deplacementPhase1Autonomous === "true") {
+        result.noMeetingPoint.youngs.push(tempYoung);
+      } else if (young.transportInfoGivenByLocal === "true") {
+        result.transportInfoGivenByLocal.youngs.push(tempYoung);
+      } else {
+        const youngMeetingPoint = meetingPoints.find((meetingPoint) => meetingPoint.meetingPointId === young.meetingPointId);
+        const youngLigneBus = ligneBus.find((ligne) => ligne._id.toString() === young.ligneId);
+        if (youngMeetingPoint) {
+          if (!result[youngLigneBus.busId]) {
+            result[youngLigneBus.busId] = {};
+            result[youngLigneBus.busId]["youngs"] = [];
+            result[youngLigneBus.busId]["ligneBus"] = [];
+            result[youngLigneBus.busId]["meetingPoint"] = [];
+          }
+          if (!result[youngLigneBus.busId]["meetingPoint"].find((meetingPoint) => meetingPoint.meetingPointId === youngMeetingPoint.meetingPointId)) {
+            result[youngLigneBus.busId]["meetingPoint"].push(youngMeetingPoint);
+          }
+          if (!result[youngLigneBus.busId]["ligneBus"].find((ligne) => ligne._id.toString() === young.ligneId)) {
+            result[youngLigneBus.busId]["ligneBus"].push(youngLigneBus);
+          }
+          result[youngLigneBus.busId]["youngs"].push(tempYoung);
+        }
+      }
+    }
+    // Transform data into array of objects before excel converts
+    const formatedRep = Object.keys(result).map((key) => {
+      let name;
+      if (key === "noMeetingPoint") name = "Autonome";
+      else if (key === "transportInfoGivenByLocal") name = "Services locaux";
+      else name = key;
+      // console.log(name, result[key]);
+      return {
+        name: name,
+        data: result[key].youngs.map((young) => {
+          const meetingPoint = young.meetingPointId && result[key].meetingPoint.find((mp) => mp._id === young.meetingPointId);
+          const ligneBus = young.ligneId && result[key].ligneBus.find((lb) => lb._id === young.ligneId);
+          // console.log(young.ligneId);
+          return {
+            _id: young._id,
+            Cohorte: young.cohort,
+            Prénom: young.firstName,
+            Nom: young.lastName,
+            "Date de naissance": formatDateFRTimezoneUTC(young.birthdateAt),
+            Sexe: translate(young.gender),
+            Email: young.email,
+            Téléphone: young.phone,
+            "Adresse postale": young.address,
+            "Code postal": young.zip,
+            Ville: young.city,
+            Département: young.department,
+            Région: young.region,
+            Statut: translate(young.statusPhase1),
+            "Prénom représentant légal 1": young.parent1FirstName,
+            "Nom représentant légal 1": young.parent1LastName,
+            "Email représentant légal 1": young.parent1Email,
+            "Téléphone représentant légal 1": young.parent1Phone,
+            "Statut représentant légal 1": translate(young.parent1Status),
+            "Prénom représentant légal 2": young.parent2FirstName,
+            "Nom représentant légal 2": young.parent2LastName,
+            "Email représentant légal 2": young.parent2Email,
+            "Téléphone représentant légal 2": young.parent2Phone,
+            "Statut représentant légal 2": translate(young.parent2Status),
+            "Id du point de rassemblement": young.meetingPointId,
+            "Adresse point de rassemblement": meetingPoint?.address,
+            "Date aller": formatDateFR(ligneBus?.departuredDate),
+            "Date retour": formatDateFR(ligneBus?.returnDate),
+          };
+        }),
+      };
+    });
+
+    const fileType = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet;charset=UTF-8";
+    const wb = XLSX.utils.book_new();
+    formatedRep.forEach((sheet) => {
+      let ws = XLSX.utils.json_to_sheet(sheet.data);
+      XLSX.utils.book_append_sheet(wb, ws, sheet.name.substring(0, 30));
+    });
+    const excelBuffer = XLSX.write(wb, { bookType: "xlsx", type: "array" });
+    const resultData = new Blob([excelBuffer], { type: fileType });
+    FileSaver.saveAs(resultData, `Listes_volontaires_par_ligne_${dayjs().format("YYYY-MM-DD_HH[h]mm[m]ss[s]")}`);
+  } catch (e) {
+    console.log(e);
+    toastr.error("Erreur !", translate(e.code));
+  }
+};
+
+async function getAllResults(index, query) {
+  const result = await API.post(`/es/${index}/export`, query);
+  if (!result.data.length) return [];
+  return result.data;
 }
